Extract surveys specs
